clList call metamethod act as iterator AND find()
To allow simpler makefile syntax, clList#s call metamethod act as
iterator (if called without parameters) and as find() (if called with
parameters).

This helps avoiding ":" for method calls in makefiles. Users should be
able to avoid fighting with "."s and ":"s in makefiles to keep the
learning curve flat.