cleanup. removing obsolete code.
There was a bug caused by the fact that source files are stored
internally only one. If i uded a source file multiple times with
different include files, the dependency tree got mixed up and
unreliable. Now source files are stored internally as often they are
used and the dependency tree works well. But this results in multiple
filetime checks for multiple used source files and the runtime for pass2
grows a little bit. Further speed optimization is nessesary.
The tools targetfile and tempfile need to be reworked ...